Move API-specific code to proper file. Actually use Validator_API class...whoops.
